requirements.txt
Pillow==5.4.1
pkg-resources==0.0.0
tqdm==4.31.1
